Creation timestamp and key expiry for shards.
Shards now have a creation timestamp and an expiry for the key.

All shards now have a timestamp denoting SystemTime::now() for the shard on
writing of the footer.

As for key expiration, local shards do not have a key expiration by default,
and keyed shards must be creating with have a key expiration time.

The shard manager will skip loading shards with an expired key, and by
default shards that are more than 7 days expired are deleted.