Fragmentation Prevention
For https://linear.app/xet/issue/XET-246/fragmentation-prevention
We use average chunks / range as a fragmentation estimator, targetting
an average of 16 chunks per range which roughly equates to 1MB per
range. This is computed over the last window of 32 ranges.

In the event of high fragmentation, we simply avoid dedupe.