Suppress errors on cache.put
The cache.put call seems to fail due to various disk errors under some
circumstances. For instance, it appears to be responsible for [this
error](https://x.com/realmrfakename/status/1915575840261513661?s=46).

If the cache.put fails the rest of the get_one_term call should still
succeed.

Logging the message as `info` rather than `error` or `warn` because
it can happen somewhat frequently (I was able to repro it with a whole
spew of errors from a single file download, repeatedly); we don't want
to litter the user's console in that case.