`upload_bytes` and pointer file "removal" (#273)
We need to support the ability for clients to upload binary data
directly to xet, so we are adding an `upload_bytes` function to
`hf_xet`.

This required some refactoring since files were baked into the upload
process. Instead of lightly changing `PointerFile` to support this,
we're taking the opportunity to refactor it out since it doesn't
conceptually make sense in the workflow we need to support.

I took a light hand to the test refactoring. IMO, the bash integration
tests should be turned into python/JS integration tests, but that seemed
too invasive a change for a small feature like this.

I'm also taking the opportunity to clean up the `hf_xet` interface a
bit. The whole pointer file concept leaks across the API, so we're
moving towards a more conceptually accurate set of data structures. We
have a `PyXetDownloadInfo` struct to support specifying the information
we need to download a file, and we have a `PyXetUploadInfo` struct to
report information on the file that was uploaded.

Note: since PyO3 doesn't have support for deprecation of the
Python->Rust interface, we need to support both the old and new types
until the next major version (IE, breaking) update to `hf_xet`.

@jgodlew, I would love your opinion on the `hf_xet` interface. Could you
take a very critical eye towards the struct and field naming for the
python data structures? Since we renaming is a hassle, we want names
that we can live with for a while.

Partially resolves XET-379.

---------

Co-authored-by: Sam Horradarn <sira.horradarn@huggingface.co>
Co-authored-by: Joseph Godlewski <jgodlewski@huggingface.co>