Connect detailed upload progress to hub (#301)
This PR connects detailed upload progress to the hub in a backwards
compatible way.

It works by testing the number of arguments and argument names on the
progress updating function. If the progress reporting function takes a
single argument, this function calls it using the old method; if it has
the appropriate arguments for detailed reporting -- `item_id,
completed_bytes, total_bytes, update_increment` -- then it calls it
using the new method. Additionally, if None is passed in, the progress
reporting is disabled.