Fixed race condition in dependency tracking. (#302)
Testing discovered two dependency tracking issues: 

The first is when a file references a new xorb multiple times in
non-contiguous locations. In this case, the logic will cause an
assertion failure when debug_assertions are enabled, though likely the
underlying logic is actually correct.

The second is that tracking which xorbs are part of a given session and
which xorbs have been uploaded previously was done by recording the xorb
hashes when they are registered for upload. However, it turns out this
needs to happen before registering the xorb as available for dedup;
otherwise a race condition could cause a xorb to be incorrectly
registered as already uploaded when in reality the xorb hash was in the
queue and not actually added to the session registry yet. This causes
the progress to be incorrectly counted as already completed.

A clean fix for this is to actually get that information directly from
the shard_manager instead of tracking it separately by returning whether
it's deduped against the session data or the cache. This fix allows
newly cut xorbs to immediately be used for dedup by other threads and
correctly tracks all the information across threads.