Optimize bg4 prediction  (#308)
This PR optimizes the bg4 prediction used in the compression scheme
analysis by bypassing the per-byte popcnt call, which can be quite slow.
This was shown to be a bottleneck; this PR speeds this operation up.

On Arm/neon, this uses an intrinsic vector popcnt instruction, and on
intel it uses a bit twiddling method to calculate the per-byte popcount
on u128s that is much more efficient than subsequent popcnt calls.