Consolidate progress reporting code to one crate (#318)
This PR consolidates the progress reporting code to a common crate
instead of having it spread between utils, data, and deduplication. This
is needed to allow the cas_client crate, which data depends on, to
depend on components of the upload completion tracking.

There should be no functionality change, just code movement.