Fixing the sdist install and adding better project information (#319)
Attempting to address the `sdist` problems and building from source.
Maturin doesn't allow for a setup like ours, so we have to workaround it
for now.

We're handling the following cases here:
1. build from source eg: `pip install
git+https://github.com/huggingface/xet-core.git@v1.1.1`. This required
the root level pyproject.toml.
2. build and release sdist. This required the github actions edit. Since
maturin replaces the root level pyproject with the one in the build
directory, we need to account for the change in working directory in the
project.
3. build wheel from hf_xet: maturin doesn't allow for `../` file
references, so we needed to link the LICENSE file, so that wheel build
succeeds and has the proper license information.