don't bump num concurrent downloads live (#326)
Removing the ability of the number of concurrent downloads to grow fixes
the error for "Too many open files" see #311 .

Tracking how many download tasks are running concurrently over the
download of a long and fragmented file showed a steady growth in the
number of allowed concurrent downloads (we grow by 1 for every
successful download without 403's). This causes the total number of
allowed concurrent downloads to reach ~1K and then we hit "Too many open
files" in my repro. Notably, my soft ulimit is 1024, and at ~1K (could
be a bit more or less, my tracking is not exactly accurate), for each
task we have 1 fd for a connection CF, so this is roughly when we reach
the ulimit.

We leave the constant number of concurrent downloads user configurable,
but constant. We can add a different live updating mechanism later.