Add incremental upload progress and total bytes to progress tracking. (#323)
This PR adds the ability for the completion tracker to report xorb
uploading incrementally, with file progress within the xorb being
updated in proportion to how many bytes of that file are in the xorb.

In addition, it also now tracks total byte completion, so we can also
report on new bytes uploaded instead of using the aggregated per-file
upload speed, which can be very confusing to interpret given that
deduplication is treated the same as uploading.

This PR contains the underlying machinery and tests; a future PR will
connect this to hf_xet.