Incremental progress on upload_xorb with retry_wrapper (#333)
This PR implements incremental progress reporting on the upload_xorb
function, reporting progress every 512KB of data uploaded.

In addition, errors are retried using the same retry policy as the other
clients. To get around Body::wrap_stream preventing retries due to
cloning failing, this PR adds a simple retry wrapper utility that allows
the entire request to be retried instead of doing it as part of the
middleware layer.