Streamline and aggregate file updates for reporting to python (#340)
With the current incremental progress updates, the amount of updates
going to python is substantial, and each has to acquire a global GIL
lock. This negatively affects the upload speed on fast connections.

This PR introduces an intermediate aggregation class that quickly
aggregates all the incoming progress updates, then sends the aggregated
update list to hf_xet once every 200 ms. With this, the thread
contention experienced by the frequent incremental updates is eliminated
while still reporting accurate progress to the user.