Added tests for session resuming +  fixed reporting issues. (#352)
This PR adds in a test suite to verify that interrupted session indeed
resume correctly, with previously uploaded xorbs correctly tracked and
reported. Along with this test suite, it fixes a reporting issue that
causes such xorbs to incorrectly be labeled in the progress tracking as
still needing to be uploaded, which results is these files appearing to
only partially complete.