Improved XORB compression method determination (#355)
Currently, to save time, we determine the optimal compression scheme by
running a detection heuristic on the first xorb encountered during an
upload session and using that to set the compression scheme for the
entire file. However, this can leave a lot on the table as it
effectively just uses the first file -- often chosen at random due to
the parallelism -- to determine the compression for the entire batch.

This PR strikes a compromise position, where the compression is
determined periodically -- specifically at the start of every contiguous
run of chunks that come from the same file or every 32 chunks
(configurable). This effectively reduces the time spent determining the
optimal compression scheme to a negligible computation burden while
still adapting to different file types and file sections within the same
repo.