Report completion speed with progress updates (#361)
Estimating completion speed turns out to be a bit annoying on the python
side; it is much easier to do it as part of the aggregation stage of the
progress updater. This PR adds two option fields that report the average
speed over the last 10 seconds (configurable) for the total completion
speed and the total transfer speed.