Move batch file uploads to FileUploadSession. (#362)
This PR moves the batch file upload code to within the
FileUploadSession. This allows us to scan all of the file sizes and
check for their existence on the start of an upload. With this, we can
set the total for the upload progress bar at the start of a session
instead of setting it dynamically as new files are added.