Make shard session resume robust to multiple sessions. (#365)
Currently, the shard merging algorithm, which is part of the shard
resume session, is not robust to shard files being deleted. This can
happen when multiple upload_files are being called, each from the same
endpoint. This PR makes this routine robust to these files disappearing.

In addition, we add a configurable environment variable,
HF_XET_SESSION_RESUME_ENABLE, that can be set to 0 / false to disable
session resume behavior. This is true by default.