Debug Symbol cleanup and instructions (#348)
A few adjustments to the debug symbol build and release process:

1. Consolidating the debug symbols into a single zip file to clean up
the release asset list and the contents of each zip.
2. Baking the platforms into the names, so we don't have heterogenous
layers of zip files.
3. Added instructions for how to use the debug symbols to the top-level
readme.

Note: for Linux, since we use the fully qualified platform name in the
symbol linking phase, this will work. Mac doesn't do any name matching
for dSYM files, it only matters that the relocation file is correct.
Windows is unchanged.

Resolves XET-582 and XET-571.