Small optimizations for chunking / upload path (#371)
This PR converts the fundamental data type in the Chunk class from
Arc<[u8]> to bytes::Bytes. The latter functions like the former, but
allows us to avoid a copy of the data on creation as conversion from a
vector is trivial. In addition, it allows creating slices by reference,
so chunks can just refer to their original data without a copy. This is
now okay, as we convert the xorbs to their compressed form quickly so
chunks do not hang around long, meaning the extra memory overhead of
this is negligible.