Switch reqwest to rustls-tls from default; use hickory-dns for dns resolution. (#378)
This PR switches reqwest to use the native-tls package, which provides a
more reliable abstraction over linux openssl implementation oddities. In
addition, it enables the trust-dns feature to embed a more robust dns
resolution path, which hopefully will fix the dns name resolution errors
such as

Now, the cas_client package exposes three features. The native-tls and
native-tls-vendored are passed on to hf_xet for building wheels, while
rustls-tls is the default.

```
# rustls-ssl embeds all ssl stuff in a rust package.  This is the most portable option, but also may not respect local 
# network configurations.   Default.
rustls-tls = ["reqwest/rustls-tls"]

# Uses native tls in the request package; this uses the native-tls package to wrap openssl, which is a more robust and portable
# way of ensuring that tls just works. 
native-tls = ["reqwest/native-tls"]

# This uses the above, but statically compiles in openssl, which makes the result more portable at the expense of 
# library size. 
native-tls-vendored = ["reqwest/native-tls-vendored"]
```

To enable this, some dependencies were moved around so that the hf_xet
package now depends directly on cas_client, which is the only package to
depend on reqwest.

---------

Co-authored-by: marked23 <@marked23>