Update shard cache default size. (#381)
After some back-of-the-envelop calculations and looking at some of our
users and what they're uploading, I think a 16GB limit on the shard
cache size is more appropriate. This effectively allows dedup against 16
TB of data while not being a huge burden relative to other aspects of
the hugging face cache.