Change download currency limit from local to global. (#385)
This PR changes the limit on the number of simultaneous range fetches
from a per-file limit to a global limit. It defaults to 128.

Currently, each TCP connection creates a socket, which uses up a file
handle. On OSX, this is limited to 256 by default, which means that
downloading multiple large files would quickly exhaust this.

This is also the cause of the dns resolver errors in
https://github.com/huggingface/xet-core/issues/373.

The tokio semaphore is fair, which means that permits are released in
the order in which they were requested. Thus this shouldn't have any
behavior change from the existing, except for the cap on the number of
simultaneous fetches.