Background loading for shards (#384)
Currently, loading all of the shards is done in a blocking manner, which
means that a large number of shards causes the call to upload_files to
take a long time to get started. This PR optimizes this path by loading
the lookup table sections of the shard directories in the background
while the chunking and file reading can get started.

It also introduces a new utility class, RwTaskLock, which provides a
RwLock-like interface around a value that can either be specified by the
value itself or by a future that resolves to the value. This makes it
easy to background tasks when values like lookup tables are held behind
an rwlock-like interface. This utility is self-contained and unit tests
are provided.