Generic retry wrapper to consolidate and streamline retry logic. (#397)
This PR consolidates the retry logic for the http connections around a
single utility, RetryWrapper, and integrates it more cleanly with the
logging and parsing logic. The goal is to replace the retry middleware
with this, which doesn't work with streaming connections.

This PR introduces this and replaces the simpler connection paths in
RemoteClient with this wrapper but leaves the previous logic intact. The
next step is to fully switch the remaining cases over to this wrapper to
clean up the code.

---------

Co-authored-by: Assaf Vayner <assaf@huggingface.co>