Updating chunk and shard cache default sizes (#406)
Updating the chunk and shard cache default size to use powers of 10
instead of powers of 2. See
https://github.com/huggingface/huggingface_hub/pull/3190#discussion_r2176840066
for background.

Note: This is my first real PR to the repo! I didn't do anything outside
of:
1. Change the constants to these new values
2. Verify the existing docs
3. Run `cargo test` and check to make sure there were no failing tests. 

If there were other steps I should've taken to validate the change, let
me know!

Sidenote: Noticed while working on this that I need to update the
existing huggingface_hub docs as
https://linear.app/xet/issue/XET-602/document-hf-xet-shard-cache-size-limit
grew out of date with the default shard cache size.