Respect XDG_CACHE_HOME and ~/ when setting cache directory. (#426)
Currently, we default the cache directory to `home_dir()/.cache`, but as
pointed out in https://github.com/huggingface/xet-core/issues/417, this
is the incorrect behavior. This PR switches this behavior to use the
[cache_dir()](https://docs.rs/dirs/latest/dirs/fn.cache_dir.html)
function to properly determine this.

The side effect of this, however, is that on OSX the cache dir will
change to the more standard `$HOME/Library/Caches/huggingface/xet`,
which means nothing in `~/.cache/huggingface/xet` will be valid anymore.

Fixes https://github.com/huggingface/xet-core/issues/417.