fix: singleflight owner task not removing Call from Group if dropped (#447)
When a singleflight `Group` is called by many tasks for a particular
key, one of these tasks is chosen as the `owner` to actually perform the
work. The other tasks are considered `waiters` and will wait until the
`owner` completes the work.

In the normal case, the `owner` runs the work, takes the result and
provides it to any `waiter` tasks. It is also the responsibility of the
`owner` to clean up the state it added to the singleflight `Group`,
namely, the `Call` record in the `Group`'s `CallMap`.

However, if the `owner` is `drop`'ed while waiting for the work to
complete (e.g. by its task being canceled), then the work will still
finish and any waiter tasks will be notified (as those actions are part
of a separately spawned task), but the state the owner added to the
`Group` is not cleaned up (i.e. there is still a lingering mapping of
key -> `Call`).

What this means is that if the `owner` is dropped before it can remove
the mapping, then the results of the work are permanently "cached" in
the singleflight `Group` and any subsequent tasks looking for the key
will always get back those results until the `Group` is dropped (usually
on process restart).

Since much of the work we use singleflight for is downloading immutable
blobs of data, "caching" the `Call` results doesn't sound that terrible
(we already cache some stuff outside of singleflight). However, this is
caching the `Result`, not just the `Ok(…)` variant, so if the work
returned an `Error`, then that becomes what is permanently cached,
which, for intermittent networking issues can cause problems.

This PR fixes the issue by adding a RAII guard struct when the `Call` is
added to the `CallMap` so that the `work()` function will remove the
`Call` when the owner exits the function (either successfully, due to
error/panic, or if it is dropped).

See the `test_dropped_owner` test for an example of how this situation
can occur.