run_and_extract_custom: remove use of explicit tokio_retry without utility (#460)
we had 1 case of using "raw" tokio_retry rather than the retry utility.
This was due to using a special custom parsing logic for chunks, rather
than built in json functionality. This PR adds a run_and_extract custom
that let's a user specify the function to parse the response body.