Drop "GaiResolverWithAbsolute" (#486)
Removes this custom DNS resolver that forces domain names to be
absolute. Using relative domain names is a valid practice in Kubernetes
clusters to configure proxy servers, e.g.
https://github.com/huggingface/huggingface_hub/issues/3323

The original [PR](https://github.com/huggingface/xet-core/pull/383) that
added this resolver was trying to resolve
https://github.com/huggingface/huggingface_hub/issues/3155 later
appeared to be a problem of user local misconfiguration.