Fix breaking build changes due to git safety checks. (#530)
Fix build break due to git safety checks. 

The recent logging PR added the automatic logging of the git repository
commit hash that was used to build the wheel. However, this requires
querying for this hash at build time, which requires running git. It
turns out that in the CI, the user checking out the repository and the
user building the wheel are different, which makes git upset. This PR
adds code to tell git everything will be ok.

---------

Co-authored-by: di <di@huggingface.co>