separating output providers, sequential output providers (#528)
This PR does a refactor of how we pass in the catch all "OutputProvider"
to the download mechanism.

It separates the download system to supporting "Sequential" and
"Seeking" operations:

- Seeking e.g. opening the file multiple times and seeking to location
-- this is the standard writing mechanism hf_xet uses today.
- Sequential e.g. opening a file once and writing data in order -- this
is to be used in a set of upcoming PR's/features to use the
parallel-download/sequential-write mechanism to support writing to
Stdout and to a channel buffer in memory.

To support an in memory channel with backpressure the Channel{Writer,
Stream, Reader} are introduced (re-introduced?) in utils. This
particularly could be useful in the mount functionality.