refactor to rewrite shard header with footer_len set to 0 (#551)
related to https://github.com/huggingface/xet-core/issues/547

When xet-core uploads the shard the header field where the footer length
is specified is set to 200 where it should be 0 according to the
specification.

Note: this value is ignored by the server today, but ideally we would
set this right since it can be useful to know if there is a footer on
the shard when reading the shard as a whole in a non-streaming fashion.