Detect local:// in endpoint string and direct to local CAS server. (#557)
For testing, we use a mock CAS server instance running in a local
directory. This is a fully functional server, but currently used only
for testing. This PR has the regular config path detect local:// as a
prefix, allowing a directory to be passed with "local://<dir>" as the
endpoint to use for testing and simulation.