deduplicate downloads for sequential output mode (#560)
This PR ensures that we only download each fetch info term once no
matter how many separate ranges it fulfills in sequential output mode.
The data is interned in memory until the last usage but is dropped after
the last reference to it is dropped.

CC @co42 I confirmed this reduces the 10GB downloaded for the 6GB file
(on my instance this wasn't faster but I was writing to EBS so that was
likely my bottleneck).