Mute chatty xet in migration service (#594)
The repo scanner migration services calls xet-core with tracing subscriber
set to logging at `INFO` level. As a result, xet-core has been too chatty for
repo scanner because we log at `INFO` level at many places as we wanted
to log more info in hf-xet to help diagnose problems.

This PR employs a build feature `elevated_information_level` which is
enabled only for hf-xet, such that the information we want to log in
hf-xet are still emitted at `INFO` level, but for others emitted at
`DEBUG` level, so not to clutter the repo scanner log.